Fix: Server was not exiting properly.
* handler was dealing with very long operations, sleeping, etc which
is of course not a handler behaviour. main is now waiting for a
semaphore which starts the clean exit. handler then just trigger
the exit process by unlocking this semaphore. For robustness,
handler will exit application if clean exit process had already
been called without any results for user.

* Clean stop process so we don't need to sleep and expect the end
of the process...

* Remove the boolean trick inside main LLVS loop to stop the process.
Simply send a SIGINT signal to the thread.