Fix: ViSP user could not compile the whole project anymore.
This is due to behaviour changes in FindVISP.cmake. Before FindVISP.cmake
was just looking for a cmake file, then deducting all include path
with it. Then it was directly including headers directory, build type
and library directory without any user explicit command. This extra
behaviour (which is not the behaviour of a package *finder*) had been
removed. An include_directories was missing in LLVS CMakeLists.txt but
since FindVISP was directly including headers, this problem was hidden.